requirements.txt
a==1.0
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.4.0
by==0.0.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
enum34==1.1.10
et-xmlfile==1.1.0
execnet==1.9.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
namedlist==1.8
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==2.0.1
pytest-xdist==2.5.0
python-dotenv==0.20.0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
webdriver-manager==3.8.1
wsproto==1.1.0
